test: add v1 address to photon e2e test (#1616)
* test: add v1 address to test_state_indexer_async_batched

* regenerate photon-api

* test: add v1 address to test_state_indexer_async_batched

* fix: correct function name and improve proof handling

* manual run for prover & photon

* photon-api regenerated

* format

---------

+++ sdk-libs/photon-api/src/models/_get_compressed_token_accounts_by_delegate_post_request_params.rs
@@ -19,7 +19,7 @@ pub struct GetCompressedTokenAccountsByDelegatePostRequestParams {
     #[serde(rename = "delegate")]
     pub delegate: String,
     #[serde(rename = "limit", default, skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i32>,
+    pub limit: Option<u16>,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "mint", default, skip_serializing_if = "Option::is_none")]
     pub mint: Option<String>,
